Fix handleEvent not being called within GameState
